Client side tests are now passing in Mocha, against the real API. client/runtests is a Node script that spins up an API instance and then runs the tests in CI (one-off) mode.
This is still a bit dicey in server mode, since the tests are all running against the same backend and tend to trample each other. The first solution that comes to mind is to have testem spin up multiple API instances and route each client to a different one, but that may be more trouble than it's worth.